Fixes BEGIN_EVENT_TABLE protection bug
I don't know if this is the right fix.
https://wiki.wxwidgets.org/Events says BEGIN_EVENT_TABLE's second
parameter should be what the first param inherits from.

We were getting this error:

error: 'const wxEventTable wxWindowBase::sm_eventTable' is protected

At the very least, this fix compiles.  At runtime, there are errors on
every click event, and the clue box doesn't work, so it's probably not
100% right.